refactor(config_reload): extract AgentComponents to its own module
config_reload.rs was 548 lines, exceeding the 500-line archlint
threshold. Move the AgentComponents struct + impl (5 methods) and
its 3 unit tests into a new app/agent_components.rs sub-module.

config_reload.rs: 548 → 386 lines.
agent_components.rs: 173 lines.
